Move rendering to separate file
