+++ tools/shell/linenoise/CMakeLists.txt
@@ -7,6 +7,7 @@ add_library(
         history.cpp
         linenoise.cpp
         linenoise-c.cpp
+        rendering.cpp
         terminal.cpp)
 set(SHELL_SOURCES
         ${SHELL_SOURCES} $<TARGET_OBJECTS:duckdb_linenoise>
